+++ modules/characters_encoding/builtin/cpp/nativecharsetBuiltin.cpp
@@ -64,7 +64,7 @@ Nelson::CharactersEncodingGateway::nativecharsetBuiltin(
             } else {
                 data = std::string();
             }
-            
+
         } else {
             data = param1.getContentAsCString();
         }